REST ConventionResource
-added delete method
-throws NoTFoundException (response status code 404) when some errors
occur

+++ src/resources/ConventionResource.java
@@ -13,7 +13,9 @@ import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
+import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
 
 import model.internal.Message;
 import model.listini.Convention;
@@ -22,6 +24,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
+import com.sun.jersey.api.NotFoundException;
+
+import service.BookingService;
 import service.ConventionService;
 import service.StructureService;
 
@@ -34,6 +39,8 @@ public class ConventionResource {
     private ConventionService conventionService = null;
     @Autowired
     private StructureService structureService = null;
+    @Autowired
+    private BookingService bookingService = null;
    
     @GET
     @Path("structure/{idStructure}")
@@ -43,7 +50,7 @@ public class ConventionResource {
         List<Convention> filteredConventions = null;
        
         filteredConventions = new ArrayList<Convention>();
-       
+        
         for(Convention each: this.getConventionService().findConventionsByIdStructure(idStructure)){           
             if(!each.getActivationCode().equals("thisconventionshouldntneverberemoved")){
                 filteredConventions.add(each);
@@ -90,17 +97,28 @@ public class ConventionResource {
         return convention;
     }
    
-    /*
+    
     @DELETE
-    @Path("structure/{idStructure}/{id}")
+    @Path("{id}")
     @Produces({MediaType.APPLICATION_JSON})   
     public  Integer delete(@PathParam("id") Integer id){
-        Integer ret = null;
+    	
+    	Integer count = 0;		
+		
+		if(this.getBookingService().countBookingsByIdConvention(id) > 0){
+			throw new NotFoundException("The convention you are trying to delete has links to one or more bookings." +
+					" Please try to delete the associated bookings before.");
+		}
+		
+		count = this.getConventionService().deleteConvention(id);
+		
+		if(count == 0){
+			throw new NotFoundException("Error: the convention has NOT been deleted");
+		}
+		
+		return count;
        
-        ret = this.getConventionService().deleteConvention(id);
-        return ret;
-       
-    }*/   
+    }   
    
     public ConventionService getConventionService() {
         return conventionService;
@@ -119,6 +137,16 @@ public class ConventionResource {
     public void setStructureService(StructureService structureService) {
         this.structureService = structureService;
     }
+
+
+	public BookingService getBookingService() {
+		return bookingService;
+	}
+
+
+	public void setBookingService(BookingService bookingService) {
+		this.bookingService = bookingService;
+	}
